+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### NOTE!!!
 ## After update to 2.x.x the plugin settings (xboxLiveId) need to be updated.
 
+## [2.2.4] - (xx.04.2022)
+## Changed
+- fixed MQTT device info
+- refactor send debug and info log
+- refactor send mqtt message
+- bump dependencies
+
 ## [2.2.2] - (09.03.2022)
 ## Changed
 - MQTT Client connection process